fixing errors

+++ src/figure.h
@@ -1,13 +1,12 @@
 #ifndef FIGURE_H
 #define FIGURE_H
 
-#include <cstddef>
 #include <string>
 
 class IFigure {
 public:
-    size_t virtual getHeight() const = 0;
-    size_t virtual getWidth() const = 0;
+    int virtual getHeight() const = 0;
+    int virtual getWidth() const = 0;
     std::string virtual print() const = 0;
     virtual ~IFigure() = default;
 };
